+++ client/src/config/myCollections/classifications.ts
@@ -1,0 +1,8 @@
+import config from '../config'
+
+export const commonClassifications: Record<string, string> = {
+  [`${config.env.dataApiBaseUrl}data/concept/f7ef5bb4-e7fb-443d-9c6b-371a23e717ec`]:
+    'Primary Name',
+  [`${config.env.dataApiBaseUrl}data/concept/ab99d278-9323-4d84-8e97-1846058fc587`]:
+    'Secondary Name',
+}