+++ client/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Added edit functionality for names ([#657](https://github.com/project-lux/lux-frontend/issues/657)).
+
 ### Changed
 
 ### Removed